+++ locations/models.py
@@ -107,7 +107,19 @@ class Locations(TranslationHelperMixin, AllinkTranslatedAutoSlugifyMixin, Transl
 
     @property
     def images(self):
-        return self.locationsimage_set.all()
+        """
+        backward compatibility:
+        either the images on the app are set
+        or we handle galleries with the gallery plugin in the header placeholder
+        """
+        try:
+            plugins = self.header_placeholder.get_plugins_list()
+        except:
+            plugins = None
+        if not plugins and self.preview_image:
+            return self.locationsimage_set.all()
+        else:
+            return None
 
     def value_has_changed_for_fields(instance, fields):
         """