+++ people/models.py
@@ -110,7 +110,19 @@ class People(TranslationHelperMixin, AllinkTranslatedAutoSlugifyMixin, Translata
 
     @property
     def images(self):
-        return self.peopleimage_set.all()
+        """
+        backward compatibility:
+        either the images on the app are set
+        or we handle galleries with the gallery plugin in the header placeholder
+        """
+        try:
+            plugins = self.header_placeholder.get_plugins_list()
+        except:
+            plugins = None
+        if not plugins and self.preview_image:
+            return self.peopleimage_set.all()
+        else:
+            return None
 
     @property
     def title(self):