+++ CHANGELOG.md
@@ -31,6 +31,9 @@ The version numbers referring to the version in [allink_core](git@github.com/all
 - people: firstname and lastname are not translated anymore.
     - migration "0019_migrate_firstname_lastname.py", including datamigration for this change can be found in "allink_apps.data_migrations". Ajust number for projects.
     - The fields old_lastname and old_firstname can be removed in a next release
+- testemonials: firstname and lastname are not translated anymore.
+    - migration "0000_testemonials_migrate_firstname_lastname.py", including datamigration for this change can be found in "allink_apps.data_migrations". Ajust number for projects.
+    - The fields old_lastname and old_firstname can be removed in a next release
 - allink_locations: text field removed and lead added. (migrate location text)
 
 ### NEW