+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The version numbers referring to the version in [allink_core](git@github.com/all
 
 ###### DATA MIGRATIONS
 
+- all apps now have logic for showing the detail gallery (if a gallery is added in header_content placeholder it will be displayed in favor of the inline_images form model.) Also the max_count of the inline is set to one! (So future Galleries have to be )
 - testemonials: firstname and lastname are not translated anymore.
     - migration "0000_testemonials_migrate_firstname_lastname.py", including datamigration for this change can be found in "allink_apps.data_migrations". Ajust number for projects.
     - The fields old_lastname and old_firstname can be removed in a next release
@@ -34,6 +35,7 @@ The version numbers referring to the version in [allink_core](git@github.com/all
 ### NEW
 
 - config: New app added in allink_apps to store project specific settings. A migrations folder (`allink_apps_migrations.config`) is necessary in every project after this version.
+
 
 ### FIXES
 - people: manager was still looking for field "lastname". is now looking for it correctly on the main table.